requirements.txt
tkinter
pytorch
librosa
soundfile
numpy